refactoring: create components
