+++ pages/[key].js
@@ -1,7 +1,9 @@
 import styled from "styled-components";
+import Head from "next/head";
 import { useRouter } from "next/router";
 import { useEffect } from "react";
-import Image from "next/image";
+import BackgroundImage from "@/components/BackgroundImage";
+import BookDetails from "@/components/BookDetails";
 
 export default function DetailPage({
   searchedBooks,
@@ -34,74 +36,18 @@ export default function DetailPage({
   }, []);
 
   return (
-    <Main>
-      <ImageContainer>
-        <StyledImage
-          src={
-            Array.isArray(currentBook.isbn)
-              ? `https://covers.openlibrary.org/b/isbn/${currentBook.isbn[0]}-L.jpg`
-              : `https://covers.openlibrary.org/b/isbn/${currentBook.isbn}-L.jpg`
-          }
-          fill
-          alt={`cover image of ${currentBook.title}`}
-        />
-      </ImageContainer>
-      <BookDetails>
-        <p>{currentBook.author_name}</p>
-        <h4>{currentBook.title}</h4>
-        <p>(first published: {currentBook.first_publish_year})</p>
-        <article className="description">
-          {!description && "no description available"}
-          {description?.value ? description.value : description}
-        </article>
-      </BookDetails>
-    </Main>
+    <>
+      <Head>
+        <title>Book Quest</title>
+      </Head>
+      <Main>
+        <BackgroundImage currentBook={currentBook} />
+        <BookDetails currentBook={currentBook} description={description} />
+      </Main>
+    </>
   );
 }
 const Main = styled.main`
   display: flex;
   justify-content: center;
-`;
-const ImageContainer = styled.div`
-  position: relative;
-  width: 100vw;
-  height: 160vw;
-`;
-
-const StyledImage = styled(Image)`
-  mask-image: linear-gradient(to bottom, rgba(0, 0, 0, 1), rgba(0, 0, 0, 0));
-  z-index: -1;
-`;
-
-const BookDetails = styled.article`
-  position: fixed;
-  top: 70%;
-  width: 90vw;
-  max-width: 700px;
-  padding: 0.5rem;
-  background-color: rgba(255, 255, 255, 0.5);
-
-  .description {
-    overflow-y: scroll;
-    ::-webkit-scrollbar {
-      width: 0.5rem;
-      height: 1em;
-      background-color: darkgrey;
-    }
-  }
-
-  @media (max-width: 799px) {
-    .description {
-      font-size: 0.77rem;
-      height: 15vh;
-    }
-  }
-
-  @media (min-width: 800px) {
-    font-size: 1.25rem;
-    .description {
-      font-size: 1rem;
-      height: 10vh;
-    }
-  }
 `;
